+++ packages/@hex-effect/infra-kysely-libsql-nats/package.json
@@ -2,7 +2,7 @@
   "name": "@hex-effect/infra-kysely-libsql-nats",
   "version": "0.1.0",
   "scripts": {
-    "build": "rm -rf dist && bun build ./index.ts --packages 'external' --outdir ./dist --target node --sourcemap='linked' && tsc --emitDeclarationOnly --outDir ./dist",
+    "build": "rm -rf dist && bun build ./src/index.ts --packages 'external' --outdir ./dist --target node --sourcemap='linked' && tsc --emitDeclarationOnly --outDir ./dist",
     "check": "tsc --noEmit",
     "test": "vitest"
   },
